chore: rename CreatePipelineButton component to PipelineActions
